% Generated by roxygen2 (4.0.1): do not edit by hand
\name{groupByOverlaps}
\alias{groupByOverlaps}
\title{Create GRangeslist from all combinations of GRanges}
\usage{
groupByOverlaps(testRanges)
}
\arguments{
\item{testRanges}{A named list of GRanges or a named GRangesList}
}
\value{
groupedGRanges A named GRangesList object.
}
\description{
Create GRangeslist from all combinations of GRanges
}
\examples{
data(ik_Example)
 gts <- groupByOverlaps(ik_Example)
}

