% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ik_Example}
\alias{ik_Example}
\title{Example Ikaros peaksets}
\usage{
data(ik_Example)
}
\value{
A list containing two GRanges objects
}
\description{
This dataset contains peaks from Ikaros ChIP by two antibodies
}
\details{
\itemize{
\item Ikpeaksets
}
}
\keyword{datasets}

