% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_import.R
\name{make_ranges}
\alias{make_ranges}
\title{Make ranges for data import
Make ranges according to specified style: use regions as-is, expand all to same width,
add a constant flank to each side (in bp), or extend by some % to each side.}
\usage{
make_ranges(testRanges, style = "region", expand = NULL, flank = NULL,
  flankUp = NULL, flankDown = NULL)
}
\description{
Make ranges for data import
Make ranges according to specified style: use regions as-is, expand all to same width,
add a constant flank to each side (in bp), or extend by some % to each side.
}

