% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{makebedtable}
\alias{makebedtable}
\title{Make HTML page for interval files or GRanges}
\usage{
makebedtable(grangesObject, name, basedirectory)
}
\arguments{
\item{grangesObject}{A GRanges or object which may be parsed by ChIPQC GetGRanges method.}

\item{name}{Name for IGV session xml}

\item{basedirectory}{Directory to write IGV xml file}
}
\description{
Creates HTML page for interval files or GRanges
}
\author{
Thomas Carroll
}

