% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{MakeIGVSampleMetadata}
\alias{MakeIGVSampleMetadata}
\title{Make sample metadata file for use with IGV.}
\usage{
MakeIGVSampleMetadata(sampleMetadata, SampleSheet, igvdirectory)
}
\arguments{
\item{sampleMetadata}{A data.frame object with metadata information for samples.
First column must contain unique sample ids.}

\item{SampleSheet}{A data.frame of file locations. First column must contain the unique sample ids.}

\item{Directory}{where igv xml files are located}
}
\description{
Creates sample metadata file for IGV from data.frames of user supplied metadata and
file locations
}
\author{
Thomas Carroll
}

