% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{MakeIGVSessionXML}
\alias{MakeIGVSessionXML}
\title{Make session xml per file}
\usage{
MakeIGVSessionXML(SampleSheet, igvdirectory, XMLname, genomeName,
  locusName = "All")
}
\arguments{
\item{SampleSheet}{A samplesheet containing file locations}

\item{igvdirectory}{Directory for IGV}

\item{XMLname}{Name for IGV session xml}

\item{IGVdirectory}{Directory to write IGV xml file}

\item{genomeName}{genome for IGV}

\item{locusName}{locus to display in igv on loading}
}
\description{
Creates session for IGV per file
}
\author{
Thomas Carroll
}

