% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{makeMotifScoreRle}
\alias{makeMotifScoreRle}
\title{Motif score as an RLElist}
\usage{
makeMotifScoreRle(pwm, regions, genome, extend, removeRand = FALSE,
  strandScore = "mean", atCentre = FALSE)
}
\arguments{
\item{pwm}{A pwn matrix object.}

\item{regions}{GRanges object to include in pwm rlelist}

\item{genome}{A BSgenome object}

\item{extend}{bps to extend regions by}

\item{removeRand}{Remove contigs with rand string}

\item{strandScore}{Method for averaging strand. Options are max, mean, sum, bothstrands}
}
\description{
Creates rlelist of pwm scores
}
\author{
Thomas Carroll
}

