% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{makeTrackHub}
\alias{makeTrackHub}
\title{Make session xml.}
\usage{
makeTrackHub(QCobject, peaksDir, bigwigDir, IGVdirectory, genome)
}
\arguments{
\item{QCobject}{A ChIPQC object}

\item{peaksDir}{Directory containing peaks}

\item{bigwigDir}{Directory containing bigwigs}

\item{IGVdirectory}{Directory to write IGV xml file}

\item{genome}{genome for IGV}
}
\description{
Creates session for IGV from ChIPQC object
}
\author{
Thomas Carroll
}

