% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{maketracktable}
\alias{maketracktable}
\title{Make HTML page for IGV sessions}
\usage{
maketracktable(fileSheet, SampleSheet, filename, basedirectory, genome)
}
\arguments{
\item{fileSheet}{A data frame, matrix or character containing sample information locations}

\item{SampleSheet}{A data frame, matrix or character containing sample information locations}

\item{igvdirectory}{Directory for IGV}

\item{filename}{Name for IGV session xml}

\item{basedirectory}{Directory to write IGV xml file}

\item{genome}{genome for IGV}
}
\description{
Creates HTML page for IGV sessions
}
\author{
Thomas Carroll
}

