% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.r
\docType{methods}
\name{c,ChIPprofile-method}
\alias{$,ChIPprofile-method}
\alias{[[,ChIPprofile,ANY,missing-method}
\alias{c,ChIPprofile-method}
\alias{cbind,ChIPprofile-method}
\alias{rbind,ChIPprofile-method}
\title{Join, subset and manipulate ChIPprofile objects}
\usage{
\S4method{c}{ChIPprofile}(x, ..., recursive = FALSE)

\S4method{rbind}{ChIPprofile}(x, ..., deparse.level = 1)

\S4method{cbind}{ChIPprofile}(x, ..., deparse.level = 1)

\S4method{[[}{ChIPprofile,ANY,missing}(x, i, j, ...)

\S4method{$}{ChIPprofile}(x, name)
}
\arguments{
\item{j}{Should be missing}
}
\value{
A ChIPprofile object
}
\description{
Join, subset and manipulate ChIPprofile objects
}
\examples{
data(chipExampleBig)
x <- c(chipExampleBig[[1]],chipExampleBig[[2]])
y <- rbind(chipExampleBig[[1]],chipExampleBig[[2]])
}

