% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{mergeChIPprofiles}
\alias{mergeChIPprofiles}
\alias{mergeChIPprofiles,ChIPprofile,ChIPprofile-method}
\alias{mergeChIPprofiles,ChIPprofile-method}
\alias{mergeChIPprofiles.ChIPprofile}
\title{Merge ChIPprofile objects}
\usage{
\S4method{mergeChIPprofiles}{ChIPprofile}(x,y)

\S4method{mergeChIPprofiles}{ChIPprofile,ChIPprofile}(x = "ChIPprofile",
  y = "ChIPprofile")
}
\arguments{
\item{x}{A ChIPprofile object}

\item{y}{A ChIPprofile object}
}
\description{
A Merge two ChIPprofiles into 1 combined ChIPprofile
}
\author{
Thomas Carroll
}

