% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{normalise}
\alias{normalise}
\alias{normalise,ChIPprofile,character,numeric-method}
\alias{normalise,ChIPprofile-method}
\alias{normalise.ChIPprofile}
\title{Normalise ChIPprofiles}
\usage{
\S4method{normalise}{ChIPprofile}(object)

\S4method{normalise}{ChIPprofile,character,numeric}(object = "ChIPprofile",
  method = "rpm", normFactors = NULL)
}
\arguments{
\item{object}{A ChIPprofile object}

\item{method}{Normalisation method}

\item{normFactors}{A numeric vector used to scale columns or rows.}
}
\description{
Various normalisation methods for ChIPprofile objects
}
\author{
Thomas Carroll
}

