% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{normaliseQuantiles}
\alias{normaliseQuantiles}
\alias{normaliseQuantiles,ChIPprofile-method}
\alias{normaliseQuantiles.ChIPprofile}
\title{Normalise quantile}
\usage{
\S4method{normaliseQuantiles}{ChIPprofile}(object)

\S4method{normaliseQuantiles}{ChIPprofile}(object = "ChIPprofile")
}
\arguments{
\item{object}{A ChIPprofile object}
}
\description{
Quantile normalisation across bins/regions.
}
\author{
Thomas Carroll
}

