% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{Ops,ChIPprofile,ChIPprofile-method}
\alias{Ops,ChIPprofile,ChIPprofile-method}
\alias{Ops,ChIPprofile,numeric-method}
\alias{Ops,numeric,ChIPprofile-method}
\alias{log,ChIPprofile-method}
\alias{log2,ChIPprofile-method}
\alias{mean,ChIPprofile-method}
\title{Arithmetic operations}
\usage{
\S4method{Ops}{ChIPprofile,ChIPprofile}(e1, e2)

\S4method{Ops}{ChIPprofile,numeric}(e1, e2)

\S4method{Ops}{numeric,ChIPprofile}(e1, e2)

\S4method{mean}{ChIPprofile}(x, ...)

\S4method{log2}{ChIPprofile}(x)

\S4method{log}{ChIPprofile}(x, base = exp(1))
}
\arguments{
\item{e1}{ChIPprofile object}

\item{e2}{ChIPprofile object}

\item{x}{objects.}

\item{...}{further arguments passed to methods.}

\item{base}{a positive or complex number: the base with respect to which
    logarithms are computed.  Defaults to \eqn{e}=\code{exp(1)}.}
}
\value{
A ChIPprofile object of result of arithmetic operation.
}
\description{
Arithmetic operations
}
\examples{
data(chipExampleBig)
chipExampleBig[[1]] + chipExampleBig[[2]]
}

