% Generated by roxygen2 (4.0.1): do not edit by hand
\name{orientBy}
\alias{orientBy}
\title{Set strand by overlapping or nearest anchor GRanges}
\usage{
orientBy(testRanges, anchorRanges)
}
\arguments{
\item{testRanges}{The GRanges object to anchor.}

\item{anchorRanges}{A GRanges object by which to anchor strand orientation.}
}
\value{
newRanges A GRanges object.
}
\description{
Set strand by overlapping or nearest anchor GRanges
}
\examples{
data(ik_Example)
strand(ik_Example[[1]]) <- "+"
anchoredGRanges <- orientBy(ik_Example[[2]],ik_Example[[1]])
}

