% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,ChIPprofile-method}
\alias{plotHeatmap.ChIPprofile}
\title{Plot heatmaps}
\usage{
\S4method{plotHeatmap}{ChIPprofile}(object,bins=100,col=heat.colors(100), 
rowScale=TRUE,orderPosition=NULL,orderBy="maxAtPosition",...)

\S4method{plotHeatmap}{ChIPprofile}(object = "ChIPprofile", bins = 100,
  col = heat.colors(100), rowScale = TRUE, orderPosition = NULL,
  orderBy = "maxAtPosition", ...)
}
\arguments{
\item{object}{A ChIPprofile object}

\item{bins}{Numeric vector of number of bins to summarise columns over (Useful for full resolution "profile" styles). 
Default is 100. Set to NULL for no binning to be performed 
Useful for comparing multiple samples of differing depths without normalisation. Default is FALSE.}

\item{col}{Colour scale to use for heatmap.}

\item{rowScale}{TRUE or FALSE. Perform row centering and scaling. Default is TRUE.}

\item{orderPosition}{Numeric vector of positions used for sorting when orderBy is set to "maxAtPosition".
May be single value specifying index for ordering or vector of numeric values where maximum and minimum index positions
specify an index range used for sorting.}

\item{orderBy}{Character specifing method for heatmap row ordering. At present only "maxAtPosition".}

\item{...}{Additional arguments passed to image()
maxAtPosition method - Order heatmap by score at index specified in orderPosition. Ordered from maximum to minimum.}
}
\value{
A matrix of values displayed in heatmap
}
\description{
A function to plot heatmaps
}
\examples{
data(chipExampleBig)
plotHeatmap(log(chipExampleBig),orderPosition=c(100:200),bins=NULL,col=topo.colors(100))
}
\author{
Thomas Carroll
}

