% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.r
\docType{data}
\name{singleGRange}
\alias{singleGRange}
\title{A single GRange}
\usage{
data(singleGRange)
}
\value{
A single GRanges used in motif coverage example/
}
\description{
This dataset contains an rlelist of motif coverage
}
\details{
\itemize{
\item singleGRange
}
}
\keyword{datasets}

