% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summitPipeline}
\alias{summitPipeline}
\title{Returns summits and summmit scores after optional fragment length prediction and read extension}
\usage{
summitPipeline(reads, peakfile, fragmentLength, readlength)
}
\arguments{
\item{peakfile}{GRanges of genomic intervals to summit.}

\item{reads}{Character vector of bamFile location or GAlignments object}

\item{fragmentLength}{Predicted or calculated fragment length. Set as NULL for auto prediction of fragment length}

\item{readlength}{Read length of alignments.}
}
\value{
Summits A GRanges object of summits and summit scores.
}
\description{
Returns summits and summmit scores after optional fragment length prediction and read extension
}

