% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{chipExampleBig}
\alias{chipExampleBig}
\title{Example ChIPprofiles}
\usage{
data(chipExampleBig)
}
\value{
A ChIPprofile object with two rows
}
\description{
This dataset contains peaks from an in-house EBF1 ChIP-seq
}
\details{
\itemize{
\item ChIPprofiles
}
}
\keyword{datasets}

