% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findconsensusRegions}
\alias{findconsensusRegions}
\title{Plot coverage of points or regions.}
\usage{
findconsensusRegions(testRanges, bamFiles = NULL, method = "majority",
  summit = "mean", resizepeak = "asw", overlap = "any",
  fragmentLength = NULL, NonPrimaryPeaks = list(withinsample = "drop",
  betweensample = "mean"))
}
\arguments{
\item{testRanges}{Named character vector of region locations}

\item{bamFile}{Named character vector of bamFile locations}

\item{method}{Method to select reproducible summits to merge.}

\item{summit}{Only mean avaialble}

\item{resizepeak}{Only asw available}

\item{overlap}{Type of overlap to consider for finding consensus sites}

\item{fragmentLength}{Predicted fragment length. Set to NULL to auto-calculate}

\item{NonPrimaryPeaks}{A list of parameters to deal with non primary peaks in consensus regions.}
}
\value{
Consensus A GRanges object of consensus summits.
}
\description{
Plot coverage of points or regions.
}

