% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ik_Profiles}
\alias{ik_Profiles}
\title{Example Ikaros signal over peaksets}
\usage{
data(ik_Profiles)
}
\value{
A ChIPprofile object
}
\description{
This dataset contains signal over peaks from Ikaros ChIP by two antibodies
}
\details{
\itemize{
\item ik_Profiles
}
}
\keyword{datasets}

