% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{c,ChIPprofile-method}
\alias{$,ChIPprofile-method}
\alias{[[,ChIPprofile,ANY,missing-method}
\alias{c,ChIPprofile-method}
\alias{cbind,ChIPprofile-method}
\alias{rbind,ChIPprofile-method}
\title{Join, subset and manipulate ChIPprofile objects}
\usage{
\S4method{c}{ChIPprofile}(x, ..., recursive = FALSE)

\S4method{rbind}{ChIPprofile}(x, ..., deparse.level = 1)

\S4method{cbind}{ChIPprofile}(x, ..., deparse.level = 1)

\S4method{[[}{ChIPprofile,ANY,missing}(x, i, j, ...)

\S4method{$}{ChIPprofile}(x, name)
}
\arguments{
\item{j}{Should be missing}

\item{...}{objects to be concatenated.}

\item{recursive}{logical.  If \code{recursive = TRUE}, the function
    recursively descends through lists (and pairlists) combining all
    their elements into a vector.}

\item{deparse.level}{
    See \code{?base::\link[base]{cbind}} for a description of
    this argument.
  }

\item{x}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{i}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link{names}} of the
    object (or for matrices/arrays, the \code{\link{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{name}{
    A literal character string or a \link{name} (possibly \link{backtick}
    quoted).  For extraction, this is normally (see under
    \sQuote{Environments}) partially matched to the \code{\link{names}}
    of the object.
  }
}
\value{
A ChIPprofile object
}
\description{
Join, subset and manipulate ChIPprofile objects
}
\examples{
data(chipExampleBig)
x <- c(chipExampleBig[[1]],chipExampleBig[[2]])
y <- rbind(chipExampleBig[[1]],chipExampleBig[[2]])
}

