% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plotRegion}
\alias{plotRegion}
\alias{plotRegion,ChIPprofile-method}
\alias{plotRegion.ChIPprofile}
\title{Plot regions}
\usage{
\S4method{plotRegion}{ChIPprofile}(object)

\S4method{plotRegion}{ChIPprofile}(object = "ChIPprofile", gts = NULL,
  plotregion = "full", outliers = NULL)
}
\arguments{
\item{object}{A ChIPprofile object}

\item{gts}{A list of character vectors and GRanges}

\item{plotregion}{region to plot}

\item{outliers}{A numeric vector of length 1 containing proportion to exclude from limits}
}
\description{
A function to plot regions
}
\author{
Thomas Carroll
}

