% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{pwmCov}
\alias{pwmCov}
\title{Example motif coverage}
\usage{
data(pwmCov)
}
\value{
A rlelist of motif coverage
}
\description{
This dataset contains an rlelist of motif coverage
}
\details{
\itemize{
\item pwmCov
}
}
\keyword{datasets}

