% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{pwmToCoverage}
\alias{pwmToCoverage}
\title{PWM hits as an RLElist}
\usage{
pwmToCoverage(pwm, genome, min = "70\%", removeRand = FALSE)
}
\arguments{
\item{pwm}{A pwn matrix object.}

\item{genome}{A BSgenome object}

\item{min}{pwm score (as percentage of maximum score) cutoff}

\item{removeRand}{Remove contigs with rand string}
}
\description{
Creates rlelist of pwm hits.
}
\author{
Thomas Carroll
}

